+++ src/main/java/com/frostwire/jlibtorrent/swig/peer_connection_handle.java
@@ -139,7 +139,7 @@ public class peer_connection_handle {
     return libtorrent_jni.peer_connection_handle_in_handshake(swigCPtr, this);
   }
 
-  public void send_buffer(String begin, int size, int flags) {
+  public void send_buffer(String begin, int size, long flags) {
     libtorrent_jni.peer_connection_handle_send_buffer__SWIG_0(swigCPtr, this, begin, size, flags);
   }
